requirements.txt
flask>=3.0
pytest>=8.0
pytest-playwright>=0.5
